requirements.txt
Pillow
tqdm
numpy
scipy
